+++ skimage/util/__init__.py
@@ -20,26 +20,27 @@ from .shape import view_as_blocks, view_as_windows
 from .unique import unique_rows
 
 
-__all__ = ['img_as_float32',
-           'img_as_float64',
-           'img_as_float',
-           'img_as_int',
-           'img_as_uint',
-           'img_as_ubyte',
-           'img_as_bool',
-           'dtype_limits',
-           'view_as_blocks',
-           'view_as_windows',
-           'slice_along_axes',
-           'crop',
-           'compare_images',
-           'map_array',
-           'montage',
-           'random_noise',
-           'regular_grid',
-           'regular_seeds',
-           'apply_parallel',
-           'invert',
-           'unique_rows',
-           'label_points',
-           ]
+__all__ = [
+    'img_as_float32',
+    'img_as_float64',
+    'img_as_float',
+    'img_as_int',
+    'img_as_uint',
+    'img_as_ubyte',
+    'img_as_bool',
+    'dtype_limits',
+    'view_as_blocks',
+    'view_as_windows',
+    'slice_along_axes',
+    'crop',
+    'compare_images',
+    'map_array',
+    'montage',
+    'random_noise',
+    'regular_grid',
+    'regular_seeds',
+    'apply_parallel',
+    'invert',
+    'unique_rows',
+    'label_points',
+]
